+++ Changes.md
@@ -19,7 +19,9 @@ Fixes
 -----
 
 - PathListingWidget : Fixed subtle bugs in the underlying Qt model, although they haven't been observed to cause problems in practice.
-- Animation : Fixed bug in `Key.setType()`. Previously it modified the value instead of the type.
+- Animation :
+  - Fixed bug in `Key.setType()`. Previously it modified the value instead of the type.
+  - Fixed crash when dragging multiple keys around in editor
 
 API
 ---
@@ -42,6 +44,12 @@ API
 - ScenePath : Added support for cancellation of calls to `isValid()` and `children()`.
 - FileSystemPath : Added support for cancellation of calls to `children()` and `property()`.
 - PathFilter : Added support for cancellation of calls to `filter()`.
+- Animation :
+  - `curve.addKey()` and `key.setTime()` now return active clashing key
+  - `curve.addKey()` now has optional remove parameter that defaults to true, if false active clashing key is not removed and becomes inactive.
+  - `key.setTime()` will not remove active clashing key, instead it becomes become inactive. inactive key at old time becomes active.
+  - Added `curve.removeInactiveKeys()` function, removes all the inactive keys parented to a curve.
+  - Added `key.isActive()` function, returns true if key is parented to curve and is the active key at its current time.
 
 Breaking Changes
 ----------------